+++ massiv/src/Data/Massiv/Array/Stencil/Convolution.hs
@@ -23,6 +23,10 @@ import GHC.Exts (inline)
 -- | Create a convolution stencil by specifying border resolution technique and
 -- an accumulator function.
 --
+-- /Note/ - Using `Data.Massiv.Array.Stencil.Unsafe.makeUnsafeConvolutionStencil` will be
+-- much faster, therefore it is recommended to switch from this function, after manual
+-- verification that the created stencil behaves as expected.
+--
 -- ==== __Examples__
 --
 -- Here is how to create a 2D horizontal Sobel Stencil:
@@ -39,7 +43,7 @@ makeConvolutionStencil
   :: (Index ix, Num e)
   => Sz ix
   -> ix
-  -> ((ix -> Value e -> Value e -> Value e) -> Value e -> Value e)
+  -> ((ix -> e -> e -> e) -> e -> e)
   -> Stencil ix e e
 makeConvolutionStencil !sz !sCenter relStencil =
   Stencil sz sInvertCenter stencil
@@ -66,7 +70,7 @@ makeConvolutionStencilFromKernel kArr = Stencil sz sInvertCenter stencil
     !szi1 = liftIndex (subtract 1) szi
     !sInvertCenter = liftIndex2 (-) szi1 sCenter
     !sCenter = liftIndex (`quot` 2) szi
-    stencil uget _ !ix = Value (ifoldlS accum 0 kArr) where
+    stencil uget _ !ix = ifoldlS accum 0 kArr where
       !ixOff = liftIndex2 (+) ix sCenter
       accum !acc !kIx !kVal = uget (liftIndex2 (-) ixOff kIx) * kVal + acc
       {-# INLINE accum #-}
@@ -76,12 +80,16 @@ makeConvolutionStencilFromKernel kArr = Stencil sz sInvertCenter stencil
 
 -- | Make a <https://en.wikipedia.org/wiki/Cross-correlation cross-correlation> stencil
 --
+-- /Note/ - Using `Data.Massiv.Array.Stencil.Unsafe.makeUnsafeCorrelationStencil` will be
+-- much faster, therefore it is recommended to switch from this function, after manual
+-- verification that the created stencil behaves as expected.
+--
 -- @since 0.1.5
 makeCorrelationStencil
   :: (Index ix, Num e)
   => Sz ix
   -> ix
-  -> ((ix -> Value e -> Value e -> Value e) -> Value e -> Value e)
+  -> ((ix -> e -> e -> e) -> e -> e)
   -> Stencil ix e e
 makeCorrelationStencil !sSz !sCenter relStencil = Stencil sSz sCenter stencil
   where
@@ -103,7 +111,7 @@ makeCorrelationStencilFromKernel kArr = Stencil sz sCenter stencil
   where
     !sz = size kArr
     !sCenter = liftIndex (`div` 2) $ unSz sz
-    stencil uget _ !ix = Value (ifoldlS accum 0 kArr) where
+    stencil uget _ !ix = ifoldlS accum 0 kArr where
       !ixOff = liftIndex2 (-) ix sCenter
       accum !acc !kIx !kVal = uget (liftIndex2 (+) ixOff kIx) * kVal + acc
       {-# INLINE accum #-}