+++ massiv/src/Data/Massiv/Array/Stencil/Internal.hs
@@ -13,7 +13,6 @@
 --
 module Data.Massiv.Array.Stencil.Internal
   ( Stencil(..)
-  , Value(..)
   , dimapStencil
   , lmapStencil
   , rmapStencil
@@ -30,104 +29,12 @@ import Data.Massiv.Core.Common
 data Stencil ix e a = Stencil
   { stencilSize   :: !(Sz ix)
   , stencilCenter :: !ix
-  , stencilFunc   :: (ix -> e) -> (ix -> Value e) -> ix -> Value a
+  , stencilFunc   :: (ix -> e) -> (ix -> e) -> ix -> a
   }
 
 
 instance Index ix => NFData (Stencil ix e a) where
   rnf (Stencil sz ix f) = sz `deepseq` ix `deepseq` f `seq` ()
-
--- | This is a simple wrapper for value of an array cell. It is used in order to improve safety of
--- `Stencil` mapping. Using various class instances, such as `Num` and `Functor` for example, make
--- it possible to manipulate the value, without having direct access to it.
-newtype Value e = Value
-  { unValue :: e
-  } deriving (Bounded)
-
-instance Functor Value where
-  fmap f (Value e) = Value (f e)
-  {-# INLINE fmap #-}
-
-instance Applicative Value where
-  pure = Value
-  {-# INLINE pure #-}
-  (<*>) (Value f) (Value e) = Value (f e)
-  {-# INLINE (<*>) #-}
-
--- | @since 0.1.5
-instance Semigroup a => Semigroup (Value a) where
-  Value a <> Value b = Value (a <> b)
-  {-# INLINE (<>) #-}
-
--- | @since 0.1.5
-instance Monoid a => Monoid (Value a) where
-  mempty = Value mempty
-  {-# INLINE mempty #-}
-  Value a `mappend` Value b = Value (a `mappend` b)
-  {-# INLINE mappend #-}
-
-instance Num e => Num (Value e) where
-  (+) = liftA2 (+)
-  {-# INLINE (+) #-}
-  (*) = liftA2 (*)
-  {-# INLINE (*) #-}
-  negate = fmap negate
-  {-# INLINE negate #-}
-  abs = fmap abs
-  {-# INLINE abs #-}
-  signum = fmap signum
-  {-# INLINE signum #-}
-  fromInteger = Value . fromInteger
-  {-# INLINE fromInteger #-}
-
-instance Fractional e => Fractional (Value e) where
-  (/) = liftA2 (/)
-  {-# INLINE (/) #-}
-  recip = fmap recip
-  {-# INLINE recip #-}
-  fromRational = pure . fromRational
-  {-# INLINE fromRational #-}
-
-instance Floating e => Floating (Value e) where
-  pi = pure pi
-  {-# INLINE pi #-}
-  exp = fmap exp
-  {-# INLINE exp #-}
-  log = fmap log
-  {-# INLINE log #-}
-  sqrt = fmap sqrt
-  {-# INLINE sqrt #-}
-  (**) = liftA2 (**)
-  {-# INLINE (**) #-}
-  logBase = liftA2 logBase
-  {-# INLINE logBase #-}
-  sin = fmap sin
-  {-# INLINE sin #-}
-  cos = fmap cos
-  {-# INLINE cos #-}
-  tan = fmap tan
-  {-# INLINE tan #-}
-  asin = fmap asin
-  {-# INLINE asin #-}
-  acos = fmap acos
-  {-# INLINE acos #-}
-  atan = fmap atan
-  {-# INLINE atan #-}
-  sinh = fmap sinh
-  {-# INLINE sinh #-}
-  cosh = fmap cosh
-  {-# INLINE cosh #-}
-  tanh = fmap tanh
-  {-# INLINE tanh #-}
-  asinh = fmap asinh
-  {-# INLINE asinh #-}
-  acosh = fmap acosh
-  {-# INLINE acosh #-}
-  atanh = fmap atanh
-  {-# INLINE atanh #-}
-
-
-
 
 instance Functor (Stencil ix e) where
   fmap = rmapStencil
@@ -142,7 +49,7 @@ instance Functor (Stencil ix e) where
 dimapStencil :: (c -> d) -> (a -> b) -> Stencil ix d a -> Stencil ix c b
 dimapStencil f g stencil@Stencil {stencilFunc = sf} = stencil {stencilFunc = sf'}
   where
-    sf' us s = Value . g . unValue . sf (f . us) (Value . f . unValue . s)
+    sf' us s = g . sf (f . us) (f . s)
     {-# INLINE sf' #-}
 {-# INLINE dimapStencil #-}
 
@@ -157,11 +64,12 @@ dimapStencil f g stencil@Stencil {stencilFunc = sf} = stencil {stencilFunc = sf'
 lmapStencil :: (c -> d) -> Stencil ix d a -> Stencil ix c a
 lmapStencil f stencil@Stencil {stencilFunc = sf} = stencil {stencilFunc = sf'}
   where
-    sf' us s = sf (f . us) (Value . f . unValue . s)
+    sf' us s = sf (f . us) (f . s)
     {-# INLINE sf' #-}
 {-# INLINE lmapStencil #-}
 
--- | A covariant map over the right most type argument. In other words a usual Functor `fmap`:
+-- | A covariant map over the right most type argument. In other words the usual `fmap`
+-- from `Functor`:
 --
 -- > fmap == rmapStencil
 --
@@ -169,31 +77,24 @@ lmapStencil f stencil@Stencil {stencilFunc = sf} = stencil {stencilFunc = sf'}
 rmapStencil :: (a -> b) -> Stencil ix e a -> Stencil ix e b
 rmapStencil f stencil@Stencil {stencilFunc = sf} = stencil {stencilFunc = sf'}
   where
-    sf' us s = Value . f . unValue . sf us s
+    sf' us s = f . sf us s
     {-# INLINE sf' #-}
 {-# INLINE rmapStencil #-}
 
 
 
--- TODO: Figure out interchange law (u <*> pure y = pure ($ y) <*> u) and issue
--- with discarding size and center. Best idea so far is to increase stencil size to
--- the maximum one and shift the center of the other stencil so that they both match
--- up. This approach would also remove requirement to validate the result
--- Stencil - both stencils are trusted, increasing the size will not affect the
--- safety.
+-- TODO: Test interchange law (u <*> pure y = pure ($ y) <*> u)
 instance Index ix => Applicative (Stencil ix e) where
-  pure a = Stencil oneSz zeroIndex (\_ _ _ -> Value a)
+  pure a = Stencil oneSz zeroIndex (\_ _ _ -> a)
   {-# INLINE pure #-}
   (<*>) (Stencil (SafeSz sSz1) sC1 f1) (Stencil (SafeSz sSz2) sC2 f2) = Stencil newSz maxCenter stF
     where
-      stF ug gV !ix = Value (unValue (f1 ug gV ix) (unValue (f2 ug gV ix)))
+      stF ug gV !ix = f1 ug gV ix (f2 ug gV ix)
       {-# INLINE stF #-}
       !newSz =
-        Sz
-          (liftIndex2
-             (+)
-             maxCenter
-             (liftIndex2 max (liftIndex2 (-) sSz1 sC1) (liftIndex2 (-) sSz2 sC2)))
+        Sz $
+        liftIndex2 (+) maxCenter $
+        liftIndex2 max (liftIndex2 (-) sSz1 sC1) (liftIndex2 (-) sSz2 sC2)
       !maxCenter = liftIndex2 max sC1 sC2
   {-# INLINE (<*>) #-}